[fix] Shim for XCTAssertNoThrow
